+++ demo/gdml/src/jsMain/kotlin/hep/dataforge/vision/gdml/demo/GdmlJsDemoApp.kt
@@ -2,6 +2,7 @@ package hep.dataforge.vision.gdml.demo
 
 import hep.dataforge.context.Global
 import hep.dataforge.vision.Application
+import hep.dataforge.vision.bootstrap.useBootstrap
 import hep.dataforge.vision.gdml.GdmlShowcase
 import hep.dataforge.vision.gdml.toVision
 import hep.dataforge.vision.solid.three.ThreePlugin
@@ -14,6 +15,8 @@ import react.dom.render
 private class GDMLDemoApp : Application {
 
     override fun start(state: Map<String, Any>) {
+        useBootstrap()
+
         val element = document.getElementById("app") ?: error("Element with id 'app' not found on page")
 
         val context = Global.context("demo") .apply{